+++ app/Http/Controllers/User/UserController.php
@@ -244,6 +244,10 @@ class UserController extends Controller
         $filename = null;
         if ($request->hasFile('image')) {
             $image = $request->file('image');
+            //Validate image file.
+            if(!GeneralFunctions::validateUserUploadImage($image)){
+                return response()->json(['error' => 'Image must be a valid image file type and size.'], 422);
+            }
             $filename = time() . '.' . $image->getClientOriginalExtension();
             Storage::disk('public')->putFileAs('images', $image, $filename);
         }
@@ -321,11 +325,15 @@ class UserController extends Controller
         $recordtype_id = RecordType::where('type', $request->recordtype)->first()->id;
 
         if ($request->hasFile('image')) {
+            $image = $request->file('image');
+            //Validate image file.
+            if(!GeneralFunctions::validateUserUploadImage($image)){
+                return response()->json(['error' => 'Image must be a valid image file type and size.'], 422);
+            }
             // Delete old image.
             if ($dataset->image_path && Storage::disk('public')->exists('images/' . $dataset->image_path)) {
                 Storage::disk('public')->delete('images/' . $dataset->image_path);
             } 
-            $image = $request->file('image');
             $filename = time() . '.' . $image->getClientOriginalExtension();
             Storage::disk('public')->putFileAs('images', $image, $filename);
             $dataset->image_path = $filename;
